backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.4
Werkzeug==3.0.1

